CHANGE: combine test files
